+++ packages/ajax-ext/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @fmcoldays/ajax-ext
 
+## 0.2.5
+
+### Patch Changes
+
+- Eliminar notificaciones pill (toast). Solo quedan popup() y confirm()
+- Updated dependencies
+  - @fmcoldays/notify@3.0.0
+
+## 0.2.4
+
+### Patch Changes
+
+- Renombrar export `toast` a `notify` y `notify()` a `popup()` en @fmcoldays/notify
+- Updated dependencies
+  - @fmcoldays/notify@2.0.0
+
+## 0.2.3
+
+### Patch Changes
+
+- Renombrar paquete de @fmcoldays/toast a @fmcoldays/notify
+- Updated dependencies
+  - @fmcoldays/notify@1.0.0
+
 ## 0.2.2
 
 ### Patch Changes