+++ packages/all/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @fmcoldays/all
 
+## 0.1.5
+
+### Patch Changes
+
+- Eliminar notificaciones pill (toast). Solo quedan popup() y confirm()
+- Updated dependencies
+  - @fmcoldays/notify@3.0.0
+  - @fmcoldays/post-action@0.3.0
+  - @fmcoldays/ajax-ext@0.2.5
+
+## 0.1.4
+
+### Patch Changes
+
+- Renombrar export `toast` a `notify` y `notify()` a `popup()` en @fmcoldays/notify
+- Updated dependencies
+  - @fmcoldays/notify@2.0.0
+  - @fmcoldays/ajax-ext@0.2.4
+  - @fmcoldays/post-action@0.2.3
+
+## 0.1.3
+
+### Patch Changes
+
+- Renombrar paquete de @fmcoldays/toast a @fmcoldays/notify
+- Updated dependencies
+  - @fmcoldays/notify@1.0.0
+  - @fmcoldays/ajax-ext@0.2.3
+  - @fmcoldays/post-action@0.2.2
+
 ## 0.1.2
 
 ### Patch Changes