+++ packages/post-action/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @fmcoldays/post-action
 
+## 0.3.0
+
+### Minor Changes
+
+- Eliminar notificaciones pill (toast). Solo quedan popup() y confirm()
+
+### Patch Changes
+
+- Updated dependencies
+  - @fmcoldays/notify@3.0.0
+
+## 0.2.3
+
+### Patch Changes
+
+- Renombrar export `toast` a `notify` y `notify()` a `popup()` en @fmcoldays/notify
+- Updated dependencies
+  - @fmcoldays/notify@2.0.0
+
+## 0.2.2
+
+### Patch Changes
+
+- Renombrar paquete de @fmcoldays/toast a @fmcoldays/notify
+- Updated dependencies
+  - @fmcoldays/notify@1.0.0
+
 ## 0.2.1
 
 ### Patch Changes